fix: yaml parsing of profile
The `yaml_document_t` always resolve nodes to:
- YAML_MAP_TAG,
- YAML_SEQ_TAG,
- YAML_STR_TAG,
due to it all checks in document for YAML_INT_TAG were removed.

Parsing of integer in `hori_yaml_int` has been changed to parse
raw integer or !!int integer.

Fix issue or traversing set / map until top of the stack and not
to the end of element.